requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
requests>=2.31.0
aiohttp>=3.9.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.0